Further updates based on discussion in PR # 21
* Add ItemMetadatMissingRequiredFieldError exception
* Rename metadata_keyname > metadata_s3_key
* Shift BaseWorkflow to base/__init__.py and remove base_workflow.py
* Add required key to metadata mapping and corresponding unit test
* Update unit tests names to include _success
