+++ dsc/item_submission.py
@@ -14,7 +14,7 @@ class ItemSubmission:
 
     dspace_metadata: dict[str, Any]
     bitstream_uris: list[str]
-    metadata_keyname: str
+    metadata_s3_key: str
     metadata_uri: str = ""
 
     def upload_dspace_metadata(self, bucket: str) -> None:
@@ -24,9 +24,7 @@ class ItemSubmission:
             bucket: The S3 bucket for uploading the item metadata file.
         """
         s3_client = S3Client()
-        s3_client.put_file(
-            json.dumps(self.dspace_metadata), bucket, self.metadata_keyname
-        )
-        metadata_uri = f"s3://{bucket}/{self.metadata_keyname}"
+        s3_client.put_file(json.dumps(self.dspace_metadata), bucket, self.metadata_s3_key)
+        metadata_uri = f"s3://{bucket}/{self.metadata_s3_key}"
         logger.info(f"Metadata uploaded to S3: {metadata_uri}")
         self.metadata_uri = metadata_uri
