+++ dsc/workflows/__init__.py
@@ -1,0 +1,10 @@
+"""dsc.workflows.
+
+All primary functions used by CLI are importable from here.
+"""
+
+from dsc.workflows.base import BaseWorkflow
+
+__all__ = [
+    "BaseWorkflow",
+]
