+++ tests/test_base_workflow.py
@@ -1,10 +1,13 @@
 import pytest
 
-from dsc.exceptions import InvalidDSpaceMetadataError
+from dsc.exceptions import (
+    InvalidDSpaceMetadataError,
+    ItemMetadatMissingRequiredFieldError,
+)
 from dsc.item_submission import ItemSubmission
 
 
-def test_base_workflow_item_submission_iter(base_workflow_instance):
+def test_base_workflow_item_submission_iter_success(base_workflow_instance):
     assert next(base_workflow_instance.item_submissions_iter()) == ItemSubmission(
         dspace_metadata={
             "metadata": [
@@ -17,12 +20,12 @@ def test_base_workflow_item_submission_iter(base_workflow_instance):
             "s3://dsc/workflow/folder/123_01.pdf",
             "s3://dsc/workflow/folder/123_02.pdf",
         ],
-        metadata_keyname="workflow/folder/123_metadata.json",
+        metadata_s3_key="workflow/folder/123_metadata.json",
         metadata_uri="",
     )
 
 
-def test_base_workflow_create_dspace_metadata(
+def test_base_workflow_create_dspace_metadata_success(
     base_workflow_instance,
     item_metadata,
 ):
@@ -33,6 +36,15 @@ def test_base_workflow_create_dspace_metadata(
             {"key": "dc.contributor", "language": None, "value": "Author 2"},
         ]
     }
+
+
+def test_base_workflow_create_dspace_metadata_required_field_missing_raises_exception(
+    base_workflow_instance,
+    item_metadata,
+):
+    item_metadata.pop("title")
+    with pytest.raises(ItemMetadatMissingRequiredFieldError):
+        base_workflow_instance.create_dspace_metadata(item_metadata)
 
 
 def test_base_workflow_validate_dspace_metadata_success(
